facetrack/requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.26.4
scikit-learn==1.3.2
scikit-image==0.22.0
streamlit==1.31.1
plotly==5.18.0
pandas==2.2.0
imutils==0.5.4
Pillow==10.2.0
